Cardコンポーネントを固定価格販売に対応
